refactor: 메서드 명 및 패키지 변경 #29
