+++ src/main/resources/lua/deduct-stocks.lua
